// frontend/src/lib/proto/masters/v1/master_pb.d.ts
export declare type Route = Message<"masters.v1.Route"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string company_id = 2;
   */
  companyId: string;

  /**
   * @generated from field: string department_id = 3;
   */
  departmentId: string;

  /**
   * @generated from field: string code = 4;
   */
  code: string;

  /**
   * @generated from field: string name = 5;
   */
  name: string;

  /**
   * @generated from field: string description = 6;
   */
  description: string;

  /**
   * @generated from field: int32 sort_order = 7;
   */
  sortOrder: number;

  /**
   * @generated from field: bool is_active = 8;
   */
  isActive: boolean;

  /**
   * @generated from field: string created_at = 9;
   */
  createdAt: string;

  /**
   * @generated from field: string updated_at = 10;
   */
  updatedAt: string;

  /**
   * @generated from field: string deleted_at = 11;
   */
  deletedAt: string;
};